debug virt-contoller: build virt-controller with debug flags
In order to run virt-controller with a debugger complilation flags must be set
in order for the compiler to avoid optimization and inlining. It is also neseesary
to  expplicily instruct the bazel go_sdk to avoid stripping.
This script  results in an image called  kubevirt/virt-controller:debug
that is pushed to the dafult bazel image registry i.e.  registry:5000

Signed-off-by: Nir Dothan <ndothan@redhat.com>